fix compatibility with Future::XS
With Future 0.49, Future objects may be implemented with Future::XS,
which doesn't use blessed hashrefs for its objects. Instead, it provides
the udata and set_udata methods for subclasses to store data.

Update the Future prereq to 0.49 to provide the udata/set_udata methods,
and use them rather than storing directly in the object.